feat(swingset): startVat counts against meterID
With this change, `startVat` computrons are now charged against the
vat's meter, if any. They are also reported to the `runPolicy`.

`startVat` includes both the swingset startup time and the userspace
`buildRootObject()` consumption. Previously, userspace could perform
arbitrary compute during `buildRootObject()` (up to the hard per-crank
limit) without consequence. For metered vats, this meant the meter was
not deducted for this early work.

By omitting this work from the runPolicy, blocks which included a vat
creation event were probably running for longer than we wanted.

Some quick experiments show that it takes 8M computrons to start a
vat, even for a fairly simple `buildRootObject()`. We still need to
experiment to find out what a good block threshold should be, but this
should give us more accurate information to work with.