chore: clean up vat-warehouse resetWorker/stopWorker names
There used to be three methods (killWorker, terminate, abandonWorker),
but really we want exactly one of two things:

* 1: make the worker stop, probably because we're terminating the vat,
     and we don't want the child process lingering
* 2: make the worker stop and reset it's snapshot/transcript, because
     we're upgrading the vat, and we want to start with a fresh+empty slate
     with new source code

The first case now uses `stopWorker`, which makes sure the child
process is gone (and is idempotent, because vat-admin might be asked
to terminate a vat which isn't currently paged-in). We can use this
when terminating a vat, but in the future we can also use it when
"pausing" a vat. In that case, some delivery will have triggered a
problem, and we want to unwind that delivery, but *not* terminate the
vat itself. To resume that vat, we'll need a worker that didn't
observe the problematic delivery, so we need to stop that worker, so
we can start a new one (from the previous snapshot/transcript).

The second case now uses `resetWorker`, which both stops the child
process, and also clears the snapshot/transcript settings. It doesn't
delete the transcript per se (to preserve history for some deep/clever
form of replay-based upgrade), but it does update the starting index,
so subsequent worker starts will not replay any of the "old"
deliveries).