Reachable durability for the AMM (#5841)
* refactor: reachable durability for the AMM

The AMM is a nested set of objects. The AMM and the pools both need to
be durable. The AMM holds its state in lexical context, and shares an
AMMState object across with all the pools. The pools access the
AMMState lexically, but the rest of their data is in their own state.

* chore: proofreading corrections

* refactor: make amm's priceAutrhorities be ephemeral

* refactor: don't store non-durables: userSeat & subscriber/publisher

* chore: final cleanups, including rename ammState to ammPowers

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>