fix(swingset): don't delete heap snapshot if it didn't change
Deliveries to a vat might not actually change the state of the vat, if
the action is read-only and GC is good at cleaning up any temporary
objects. When the vat runs inside an `xsnap` worker, this can result
in the following heap snapshot being identical to the previous one.

`vatKeeper.saveSnapshot()` had a bug in which it always marked the
previous snapshotID as unused, even if the new snapshotID was the same
as the old one. The snapStore handled this correctly, but vatKeeper
was telling it to `prepareToDelete` by mistake.

Snapshots and the kvStore are (unfortunately) in separate databases,
with separate commit points. To avoid losing data if the application
halts after kvStore commits but before snapStore commits, we defer
deletion of unused snapshots until the `snapStore.commit()`. The
snapStore maintains an in-RAM `Set` of unreferenced snapshot IDs to
remember what needs deletion.

The vatKeeper bug manifested as a crash after the following sequence
of events:

* 1: a delivery is made that does not change the vat state (snapshotA)
  * the bug tells snapStore to add snapshotA to the deletion list
* 2: `snapStore.commit()` runs
  * snapshotA is deleted, oops
* 3: a second delivery arrives, resulting in snapshotB
  * vatKeeper tells snapStore to delete snapshotA again
* 4: `snapStore.commit()` runs again
  * snapStore attempts to re-delete the snapshotA file
  * that results in ENOENT and a crash

We only create a snapshot for every 200 cranks, so there could be a
considerable delay between events 2 and 3, allowing the crash to
happen well after the actual problem. And until recently, xsnap was
not creating deterministic snapshots, masking the problem.

The fix is to change `vatKeeper` to not assume the new snapshot is
different than the old one: if they are the same, don't mark the old
one as unused.

closes #5901