fix(zoe): fix fakePriceAuthority event ordering
The "priceAuthority quoteAtTime" test was designed to exercise the API
that waits for a particular time to arrive, then emits a price quote
for that timestamp.

Previously, the fakePriceAuthority used two uncoordinated timer
calls. The authority used `makeRepeater()` to periodically update
`currentPriceIndex` and `latestTick`. And the `quoteAtTime()` API
method used its own `setWakeup()`, to wait for a time to arrive, then
sample `currentPriceIndex` to build the quote.

The relative firing order of these nominally-simultaneous wakeup
events is not specified by the TimerService API, and apparently
changed in the new implementation. (Now that Zoe's manualTimer is
built on top of the real vat-timer implementation, the opportunity for
future divergence should be reduced).

With the old implementation, the `quoteAtTime(3n)` alarm was firing
first. It sampled `currentPriceIndex` before the `makeRepeater` wakeup
could change it. So it built a quote from the price for `time=2n`,
where `currentPriceIndex = 1` (i.e. price=55) and a timestamp of 3n.

The new implementation of fakePriceAuthority removes the uncoordinated
`setWakeup`. When `quoteAtTime` is given a time in the future, it adds
the timestamp and a `resolve` callback to a list named
`timeClients`. Then the single `makeRepeater` handler checks this list
and fires anything no longer in the future, *after* it has updated
`currentPriceIndex`.

In addition, this implementation starts its `makeRepeater()` at time
1, not time 0. The new manualTimer responds to `delay=0` by firing the
repeater immediately, whereas the old one would only fire during a
`tick()` call. This caused the new implementation to increment
`currentPriceIndex` one time too many, as it was counting wakeups, not
the changes to the timestamp received by the wakeups. As an extra
defense, the `startTicker` handler was changed to ignore a wakeup with
`time === 0`.

When run against the new implementation, the unit test sees the
price=55 happen at time `2n`, not `3n`, which I think was the original
intent.

A similar issue caused `test-callSpread.js` to need a different option
expiration time.