test(xsnap): disable timestamp test until it tolerates fast CI hosts
This test gets a false failure if the host is fast enough to cause two
events to occur within the same microsecond. The ticket describes a
weaker ordering assertion to apply that would confirm we get
timestamps and they they appear in mostly the right order, while not
failing if the CI host is too fast. I'm disabling this test until I
get around to implementing that scheme.

refs #5951