fix(swing-store): use sqlite transactions for streamStore operations
The 'streamStore' component of swing-store is used to hold the
append-only delivery transcript for each vat. It stores each record in
a SQLite table.

Previously, streamStore used SQLite's auto-commit API: every write was
committed immediately. This causes a significant performance hit,
especially on slower disks, because a DB commit must do 3-ish
`fsync()` calls. SSDs are fast, but we observed 12.8ms stalls on the
standard disks of a popular cloud CPU provider, for each call to
writeStreamItem().

For the SwingSet client, each vat delivery triggered this stall, and a
full chain block could have hundreds of deliveries, leading to
multiple seconds of delay.

This patch changes streamStore to create an explicit DB transaction
instead of relying on automatic ones. SQLite will not commit these
transactions automatically, so the patch also introduces internal
`streamStore.commit()` and `.close()` methods.

Clients should continue to use `swingStore.commit()` and
`swingStore.close()` as usual; they will invoke the internal
streamStore methods as appropriate. The three patterns are:

* `commit()`: commit changes, keep using the DB
* `commit(); close();`: commit changes, close the DB
* `close();`: abandon the changes, close the DB

The streamStore is committed before the kvStore (LMDB) commit. As
before, we rely upon clients recording the startPos/endPos of the
stream into the kvStore, to safely tolerate a crash just after the
streamStore.commit but before the kvStore.commit. The next restart
after the crash will contain the spurious stream item in streamStore,
but the old startPos/endPos in kvStore, so clients will effectively
ignore the leftover item.

closes #6056