fix(deploy-script-support): Use versioned bundleSource
This change shifts from using the endowed bundleSource from agoric-cli
to using the version of bundleSource in the contract's lockfile.

Veryfing the integrity of a contract requires that the contract be
bundled consistently and consequently have a consistent hash for
verifying its integrity.
So, the version of bundleSource must be pinned in the contract's
lockfile and cannot vary based on the version of the agoric-cli that
happens to be installed on the developers' systems.

By having contracts pin a version of deploy-script-support,
they will have a consistent version of bundleSource.