fix(swingset): only preload maxVatsOnline/2 vats
`runtimeOptions.warehousePolicy.maxVatsOnline` is an option which
limits the number of xsnap workers the kernel is willing to run at the
same time. It defaults to 50, which means when a message is sent to
the 51st vat, some other worker (the least-recently used one) is
evicted/killed (aka "paged out") to make room for the worker being
started ("paged in"). This limits the amount of RAM and processes we
consume from the host OS (usually Linux).

Previously, when the kernel starts, it would preload all vats, even if
there were more than `maxVatsOnline`, which had the pathological
behavior of 1: paging vats in only to page them back out again before
ever talking to them, and 2: leaving the most recent *defined* vats
online, rather than the most likely to be *used* vats.

In the Agoric PSM chain, we'll create a new vat for each governance
vote, and those vote vats don't get any traffic once the governance
issue is complete, but for various reasons we don't terminate
them. After 30-ish votes, there will be enough dynamic vats for the
early static vats (like bootstrap and zoe) to be evicted during kernel
startup, which wastes time and increases latency at some unpredictable
point in the future when new Zoe/contract messages arrive.

This changes the preload process to only load `maxVatsOnline / 2`
vats, not all of them. It starts with the static vats (in
lexicographic order, unfortunately, not numberic), and then goes
through all the dynamic vats (in creation order), stopping after we
hit the limit. This should avoid loading the useless
done-but-not-terminated voting vats, and should leave room in the LRU
queue so the next message that arrives to a still-offline worker won't
evict a useful one right away.

In the long run, we might want to persist the LRU order across a
reboot, so any evictions that *do* happen are less likely to hit a vat
which is probably going to get paged in again soon. We might preload
everything in the LRU (restoring the same set of workers that the
earlier incarnation had running), or only a subset at the top (under
the assumption that the bottom is unlikely to be paged in again any
time soon, and it's better to save time/memory by not loading them).

This patch also changes test-warehouse.js to be insensitive to the new
vat-*.js files added to its directory (for the new test-preload.js).

closes #6433