test: Add Node.js 18.x to workflows
Node.js 14.x is retained as the default, just as it was before,
considering it is Maintenance LTS set to EOL next year (2023-04-30).

This change adds Node.js 18.x in all the places where workflows
previously were building with Node.js 16.x - either as addition
to it where multiple version are provided, or as an upgrade where
only one version is used.

This finishes up the bulk of #6489, however manual testing of
compatibility should be conducted further before that.