test(solo): Fix Websocket for IPv6-native hosts
With Node.js 17+ defaulting to IPv6, hosts that have native IPv6 stacks
might have resolved `localhost` to `::1`. (See: nodejs/node#39987) This
would generally be fine for most applications, but seems here listening
only happens on the IPv4 stack. This was not observed as an issue on
Node.js <= 16, because it prioritized IPv4 addresses historically.

This commit replaces `localhost` with `127.0.0.1` because there is a
`connections.json` which gets written the IPv4 address as well. At a
later time, the package should theoretically be re-checked, to better
handle (and use) the IPv6 stack.