fix(swingset): always delete vom.rc key when count=0
Previously, in liveslots/virtualReferences.js, when the refcount
dropped to 0, we didn't delete the vom.rc key in the one case where
the baseref was for a virtual object. We couldn't remember why, and
this causes a DB key to be left around for an indefinite time. So this
changes `setRefCount()` to unconditionally delete the key when it
reaches zero.