fix: remove incomplete device-mock method getBundleIDByName
The fake BundleID it returns doesn't correlate with anything else it
returns, so it would be better (easier to debug) if any test that
ended up provoking it would just fail (with a "no such method" error),
rather than get an ID that wasn't known to e.g. the fake
VatAdminService.