fix(liveslots): use Map for vrefStatus, not object
In virtualReferences.js, the updateReferenceCounts() function
was (ab)using an Object to do a Map's job. We generally avoid this
pattern: while convenient, it's far too easy to accidentally think
that properties from the prototype are actually members of the
pseudo-Map (e.g. imagine if we somehow had a vref named "toString").

The fix is to just replace the object with a proper Map. It's probably
faster too.